backend/pyproject.toml
[tool.poetry]
name = "graftai-backend"
version = "1.0.0"
description = "Backend for GraftAI — AI-powered scheduling platform"
authors = ["GraftAI Team"]

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastapi = ">=0.110.0"
uvicorn = {version = ">=0.29.0", extras = ["standard"]}
pydantic = {version = ">=2.10,<3", extras = ["email"]}
bleach = ">=6.1.0"
pytz = ">=2024.1"
aiosqlite = ">=0.20.0"
asyncpg = ">=0.29.0"
sqlalchemy = {version = ">=2.0.0", extras = ["asyncio"]}
alembic = ">=1.13.0"
redis = ">=5.0.0"
pyjwt = ">=2.10.0"
cryptography = ">=43.0.0"
argon2-cffi = ">=23.1.0"
langchain = ">=0.2.0,<1.3"
langchain-openai = ">=0.1.0,<2"
langchain-community = ">=0.2.0,<1"
openai = ">=1.0.0,<3"
pinecone = ">=5.0.0"
langchain-pinecone = ">=0.2.13"
pyyaml = ">=6.0.1,<7.0"
python-dotenv = ">=1.0.1"
httpx = ">=0.28.1"
authlib = ">=1.2.0"
pyotp = ">=2.9.0"
msal = ">=1.31.0"
google-auth = ">=2.34.0"
google-auth-oauthlib = ">=1.2.1"
google-api-python-client = ">=2.142.0"
groq = ">=0.4.0"
python-multipart = ">=0.0.18"
bcrypt = ">=4.0.0"
arq = ">=0.25.0"
boto3 = ">=1.34.0"
jinja2 = ">=3.1.0"
gunicorn = ">=21.2.0"
sentry-sdk = {version = ">=2.10.0", extras = ["fastapi", "sqlalchemy", "redis", "arq"]}
stripe = ">=11.0.0"
razorpay = ">=1.4.2"
cachetools = ">=5.5.0"
msgpack = ">=1.1.0"
python3-saml = ">=1.16.0"
resend = ">=2.4.0"
uvloop = {version = ">=0.20.0", markers = "sys_platform != 'win32'"}

[tool.poetry.dev-dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.25.0"
black = "^24.0.0"

[tool.ruff]

[tool.ruff.lint]
ignore = ["E402", "E701", "E712", "F841"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"